requirements.txt
streamlit
scikit-learn
nltk
python-docx
PyPDF2
numpy
scipy
pandas